triangle_lr_cycle bug fix

+++ train.py
@@ -11,7 +11,7 @@ def adam_sgdr_one_cycle(total_steps: int, lr: float = 0.001):
 
 
 def triangle_lr_one_cycle(lr: float, step, total_steps: int, warmup_steps: int):
-    warmup_sched = lambda: step * lr / warmup_steps
+    warmup_sched = lambda: tf.cast(step, tf.float32) * lr / warmup_steps
     decay_sched = lambda: (total_steps - step) * lr / (total_steps - warmup_steps)
     return tf.cond(tf.less_equal(step, warmup_steps), warmup_sched, decay_sched)
